(Feature)|Create Launch storyboard and launch transition animation.
Summary:
Adds the launch storyboard and launchviewcontroller, and the application view controller.  The AppViewController is a container controller that owns control over the main state of the app.  Currently, it is just being used for transitioning from the launch view animations to the main app.  In the future, it can be used to transition from a login workflow to the main app and back.